Move model builder extension methods
